backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymupdf==1.23.15
python-docx==1.0.1
nltk==3.8.1
spacy==3.7.2
openai==1.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
joblib==1.3.2
